[PGPRO-11599] Fix wrong results returned when order_by_attach=TRUE. (#164)
With order_by_attach=TRUE, the RUM index arranges the entries in the posting
lists (for keys that have additional information) in order of sorting by
additional information. For the remaining keys, the records are sorted by tid.
The scanGetItemRegular() function assumes that all posting lists have the same
sorting, so it was not suitable for sorting in this case.

In order to fix this, the appropriate TIDs for keys that are ordered by tid are
assembled in advance into a bitmap. For keys ordered in an alternative order, a
routine scan is performed, after which the appropriate tid is checked in the
bitmap.

1. Static functions from TIDBitmap were added to RUM as RumTIDBitmap.
    A function has been added to search for tid inside RumTIDBitmap.
2. Added a test of scanning with keys that are ordered differently.
3. Fixed CI crashes at hardcore level.

---------

Co-authored-by: Arseny Kositsyn <a.kositsyn@postgrespro.ru>